crash fix again

+++ src/main/java/com/github/alexthe666/iceandfire/entity/EntityHippocampus.java
@@ -265,15 +265,15 @@ public class EntityHippocampus extends EntityTameable implements IAnimatedEntity
     }
 
     public boolean up() {
-        return (dataManager.get(CONTROL_STATE).byteValue() & 1) == 1;
+        return (Byte.valueOf(dataManager.get(CONTROL_STATE).byteValue()) & 1) == 1;
     }
 
     public boolean down() {
-        return (dataManager.get(CONTROL_STATE).byteValue() >> 1 & 1) == 1;
+        return (Byte.valueOf(dataManager.get(CONTROL_STATE).byteValue()) >> 1 & 1) == 1;
     }
 
     public boolean dismount() {
-        return (dataManager.get(CONTROL_STATE).byteValue() >> 2 & 1) == 1;
+        return (Byte.valueOf(dataManager.get(CONTROL_STATE).byteValue()) >> 2 & 1) == 1;
     }
